+++ DmpWorkflow/scripts/dampe_cli_create_new_job.py
@@ -30,7 +30,7 @@ def main(args=None):
     xmlFile = opts["xml"]
     t_type = opts['type']
     n_instances = opts["Ninstances"]
-    res = requests.post("%s/job/"%DAMPE_WORKFLOW_URL,
+    res = requests.post("%s/job/" % DAMPE_WORKFLOW_URL,
                         data={"taskname": taskName, "type": t_type, "n_instance": n_instances},
                         file={'job_description', open(xmlFile, "rb")})
     res.raise_for_status()
